// src/payload-types.ts
export interface User {
  id: string;
  role?: ('admin' | 'user') | null;
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  _verified?: boolean | null;
  _verificationToken?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  password?: string | null;
}
export interface Product {
  id: string;
  user?: (string | null) | User;
  name: string;
  description?: string | null;
  price: number;
  category: 'icons' | 'ui_kits' | 'Exclusivité';
  forWho: 'Femme' | 'Homme' | 'Mixte';
  approvedForSale?: ('pending' | 'approved' | 'Denied') | null;
  priceId?: string | null;
  stripeId?: string | null;
  product_files: string | ProductFile;
  Product_3D_GLTF: string | ThreeDObjectGltf;
  Product_3D_BIN: string | ThreeDObjectBin;
  product_3D_File_textures: (string | ThreeDObjectTexture)[];
  images: (string | Image)[];
  threeDObject_zip: string | ThreeDObjectZip;
  updatedAt: string;
  createdAt: string;
}
export interface ProductFile {
  id: string;
  user?: (string | null) | User;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
}
export interface ThreeDObjectGltf {
  id: string;
  user?: (string | null) | User;
  product?: (string | Product)[] | null;
  threeDObject_bin?: (string | null) | ThreeDObjectBin;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
}
export interface ThreeDObjectBin {
  id: string;
  user?: (string | null) | User;
  product?: (string | Product)[] | null;
  threeDObject_gltf?: (string | null) | ThreeDObjectGltf;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
}
export interface ThreeDObjectTexture {
  id: string;
  user?: (string | null) | User;
  product?: (string | Product)[] | null;
  threeDObject_bin?: (string | null) | ThreeDObjectBin;
  threeDObject_gltf?: (string | null) | ThreeDObjectGltf;
  prefix?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
}
export interface Image {
  id: string;
  user?: (string | null) | User;
  product?: (string | Product)[] | null;
  prefix?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
}
export interface ThreeDObjectZip {
  id: string;
  user?: (string | null) | User;
  product?: (string | Product)[] | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
}